add hnswlib option
